Move schema to utils
