+++ esmerald/utils/schema.py
